+++ src/ui/util/routes.ts
@@ -160,6 +160,7 @@ const routes = () => {
 		"/l/:lid/frivolities/colleges": genPage("colleges"),
 		"/l/:lid/frivolities/countries": genPage("countries"),
 		"/l/:lid/frivolities/draft_classes": genPage("frivolitiesDraftClasses"),
+		"/l/:lid/frivolities/jersey_numbers": genPage("frivolitiesJerseyNumbers"),
 		"/l/:lid/frivolities/teams/:type": genPage("frivolitiesTeamSeasons"),
 		"/l/:lid/frivolities/most/:type": genPage("most"),
 		"/l/:lid/frivolities/relatives": genPage("relatives"),