+++ src/ui/views/Colleges.tsx
@@ -5,7 +5,7 @@ import { DataTable, PlayerNameLabels } from "../components";
 import type { View } from "../../common/types";
 import { frivolitiesMenu } from "./Frivolities";
 
-export const genView = (type: "college" | "country") => {
+export const genView = (type: "college" | "country" | "jerseyNumbers") => {
 	return ({
 		challengeNoRatings,
 		infos,
@@ -14,7 +14,12 @@ export const genView = (type: "college" | "country") => {
 		valueStat,
 	}: View<"colleges">) => {
 		useTitleBar({
-			title: type === "college" ? "Colleges" : "Countries",
+			title:
+				type === "college"
+					? "Colleges"
+					: type === "country"
+					? "Countries"
+					: "Jersey Numbers",
 			customMenu: frivolitiesMenu,
 		});
 
@@ -34,7 +39,11 @@ export const genView = (type: "college" | "country") => {
 		];
 
 		const cols = getCols(
-			type === "college" ? "College" : "Country",
+			type === "college"
+				? "College"
+				: type === "country"
+				? "Country"
+				: "stat:jerseyNumber",
 			"# Players",
 			"# Active",
 			"# HoF",
@@ -80,6 +89,7 @@ export const genView = (type: "college" | "country") => {
 						classNames: {
 							"table-danger": p.hof,
 							"table-success": p.retiredYear === Infinity,
+							"table-info": p.statsTids.includes(userTid),
 						},
 					},
 					p.ratings[p.ratings.length - 1].pos,
@@ -110,9 +120,10 @@ export const genView = (type: "college" | "country") => {
 		return (
 			<>
 				<p>
-					Active players are{" "}
-					<span className="text-success">highlighted in green</span>. Hall of
-					Famers are <span className="text-danger">highlighted in red</span>.
+					Players who have played for your team are{" "}
+					<span className="text-info">highlighted in blue</span>. Active players
+					are <span className="text-success">highlighted in green</span>. Hall
+					of Famers are <span className="text-danger">highlighted in red</span>.
 				</p>
 				<DataTable
 					cols={cols}