+++ src/ui/views/FrivolitiesJerseyNumbers.tsx
@@ -1,0 +1,3 @@
+import { genView } from "./Colleges";
+
+export default genView("jerseyNumbers");